Joy Mech Fight detector
Untested on FC2 as variables don't show up in test mode for NES games, but these addresses are identical between MAME and other emulators.

Extra information:
Only Player 1 has control on the main menu. To prevent potential abuse leading to unfair wins and justifiable ragequits, check if both 0x536 and 0x636 are 0 when a fight starts. These addresses determine whether an AI or human is in control for P1 and P2, respectively. The community has never needed this, but I want to make sure it never happens.
Draws do not affect anything here. No win flags are set off and the match resets (no changing characters). I'm going to add it anyway in case it's needed or helpful for data like rank points or match count: 0x75 = 16 ($10)